requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
streamlit>=1.22.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=1.10.7
